Improve girder-mount command.
Add diskcache optionally to mount.  Add flatten option for girder mount.

This allows some caching at the fuse mount level which can reduce
traffic to non-filesystem assetstores (such as S3).  This substantially
helps speeds when using inefficient file readers on such assetstores
(e.g., openslide reading Hamamatsu ndpi files).

The resource path flatten option allows single file items to appear as
/collection/<collection name>/<folder name>/<file name> rather than
/collection/<collection name>/<folder name>/<item name>/<file name>.
This is optional and can selectively be applied to the paths in the
mount.  The virtue of this is that many file readers that work with file
formats that use multiple files expect these files to be in a particular
location relative to one another.  For imported files on filesystem
assetstores where the mount isn't used, this isn't a problem.  But for
uploaded files or files on non-filesystem assetstores, the item/file
path structure usually doesn't work for these readers.